pyproject.toml
[project]
name = "databases"
version = "0.1.0"
description = ""
authors = [
    {name = "Nurlan, nurlan.grachyov@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy[asyncio] (>=2.0.44,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "lxml (>=6.0.1,<7.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "xlrd (>=2.0.2,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "aiohttp (>=3.13.1,<4.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
]
packages = [{ include = "*" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
